102_agent_project/requirements.txt
langchain-community
langchain-core
langchain-openai
langgraph
faiss-cpu
sentence-transformers
ragas
streamlit
python-dotenv
datasets
